feat: add execution analytics
- also ensure JobID is called job_id in all events for easier sql
